refactor(dashboard): split redteam_views, share transcript renderer, drop dead code
Part A: split redteam_views.py (970 lines) into redteam_charts.py
(breakdown + heatmap + their helpers) and redteam_transcripts.py
(conversation viewer + disagreement viewer). redteam_views.py is now a
thin route-wiring module that re-exports register_redteam_view_routes.

Part B: add render_message_list(messages, *, role_labels, class_prefix) to
view.py as the shared flat-message-list renderer; sim_views now uses it
(with pre-normalised content via coerce_content_text) instead of the
equivalent inline loop.

Part C: delete sim_views._meta (documented unused); remove the dead
filter_qs param/branch from view.download_sidebar and update its only
caller in app.py; collapse the unreachable else-branch in app.report_view
(both surfaces are in FILTERS); delete app._resolve_and_load (defined but
never called).
